requirements.txt
pandas
streamlit
plotly
azure-storage-file-datalake
python-dotenv